READY: dm: add support to directly boot to a mapped device
Add a dm-mod.create= kernel module parameter.
It allows device-mapper targets to be configured at boot time for use early
in the boot process (as the root device or otherwise).

Signed-off-by: Will Drewry <wad@chromium.org>
Signed-off-by: Kees Cook <keescook@chromium.org>
[rework to use dm_ioctl calls]
Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
[refactored for upstream]
Signed-off-by: Helen Koike <helen.koike@collabora.com>

Series-to: dm-devel@redhat.com
Series-cc: wad@chromium.org, keescook@chromium.org, snitzer@redhat.com, linux-doc@vger.kernel.org, richard.weinberger@gmail.com, linux-kernel@vger.kernel.org, linux-lvm@redhat.com, enric.balletbo@collabora.com, kernel@collabora.com, agk@redhat.com
Series-version: 11
Series-changes: 11
 - Configure the device directly instead of performing in IOCTL (this
 removed a lot of parsing code)
 - Just enable the targets that were tested.
 - Simplify/refactor parsing, as a consequence, escaping characters is not
 allowed (but it wans't properly used in the previous version anyway)
 - don't use sscanf, the size wans't being limited and we could have a
 buffer overflow.
 - change constrained targets list
 - remove code from init/
 - use a module parameter instead of a kernel comand line parameter in
 init/
 - rename dm-boot to dm-init

Series-changes: 10
 - https://lore.kernel.org/patchwork/project/lkml/list/?series=371523
 - new file: drivers/md/dm-boot.c
 - most of the parsing code was moved from init/do_mounts_dm.c to drivers/md/dm-boot.c
 - parsing code was in essence replaced by the concise parser from dmsetup
 _create_concise function:
https://sourceware.org/git/?p=lvm2.git;a=blob;f=libdm/dm-tools/dmsetup.c;h=835fdcdc75e8f0f0f7c4ed46cc9788a6616f58b8;hb=7498f8383397a93db95655ca227257836cbcac82#l1265
 the main reason is that this code is already being used/tested by dmsetup, so
 we can have some level of confidence that it works as expected. Besides this,
 it also looks more efficient.
 - Not all targets are allowed to be used by dm=, as pointed previously, there
 are some risks in creating a mapped device without some validation from
 userspace (see documentation from the patch listing which targets are allowed).
 - Instead of using a simple singly linked list (for devices and tables), use
 the struct list_head. This occupies unnecessary space in the code, but it makes
 the code cleaner and easier to read and less prone to silly errors.
 - Documentation and comments were reviewed and refactored, e.g.:
	* "is to possible" was removed
	* s/specified as a simple string/specified as a string/
 - Added docs above __align function, make it clear that the second parameter @a
 must be a power of two.
 - Clean ups: removal of unnecessary includes, macros, variables, some redundant
 checks and warnings.
 - when calling ioctls, the code was allocating and freeing the same structure
 a couple of times. So instead of executing kzalloc/kfree 3 times, execute
 kmalloc once and reuse the structure after a memset, then finally kfree it once.
 - update commit message

Series-changes: 9
 - https://www.redhat.com/archives/linux-lvm/2018-September/msg00016.html
 - Add minor number to make it compatible with dmsetup concise format

Series-changes: 8
 - https://www.redhat.com/archives/linux-lvm/2017-May/msg00055.html
 - Fix build error due commit
    e516db4f67 (dm ioctl: add a new DM_DEV_ARM_POLL ioctl)

Series-changes: 7
 - http://lkml.iu.edu/hypermail/linux/kernel/1705.2/02657.html
 - Add a new function to issue the equivalent of a DM ioctl programatically.
 - Use the new ioctl interface to create the devices.
 - Use a comma-delimited and semi-colon delimited dmsetup-like commands.

Series-changes: 6
 - https://www.redhat.com/archives/dm-devel/2017-April/msg00316.html

Series-changes: 5
 - https://www.redhat.com/archives/dm-devel/2016-February/msg00112.html

Commit-notes:
Hello,

As mentioned in previous discussions, the need to create an initramfs adds
another layer of complexity when performing tests or preparing a BSP for
a board, I particularly prefer not having an initramfs and it is always
a bit painful when I need to take care of this step and it also occupies
space.
Besides my particular preference, this is also useful for Android and
Chrome OS that don't use initramfs.
A practical example as mentioned by Kees is that Chrome OS has a limited amount
of storage available for the boot image as it is covered by the static root of
trust signature.

So, as mentioned in the previous version, this patchset allows dm
devices to be configured in the kernel command line parameter for
use early in the boot process without an initramfs.

One of the main difference in this version is that instead of using "dm=",
I'm using a parameter in the module "dm-mod.create=" (thanks Ezequiel for
the idea).

The syntax used in the boot param is based on the concise format from the dmsetup
tool as described in its man page http://man7.org/linux/man-pages/man8/dmsetup.8.html#CONCISE_FORMAT

Which is:
	dm-mod.create=<name>,<uuid>,<minor>,<flags>,<table>[,<table>+][;<name>,<uuid>,<minor>,<flags>,<table>[,<table>+]+]

Where,
	<name>		::= The device name.
	<uuid>		::= xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx | ""
	<minor>		::= The device minor number | ""
	<flags>		::= "ro" | "rw"
	<table>		::= <start_sector> <num_sectors> <target_type> <target_args>
	<target_type>	::= "verity" | "linear" | ...

Example, the following could be added in the boot parameters.
dm-mod.create="lroot,,,rw, 0 4096 linear 98:16 0, 4096 4096 linear 98:32 0" root=/dev/dm-0

Please check the patch with the documentation on the format.

The idea to make it compatible with the dmsetup concise format is to make it
easier for users, allowing just copy & paste from the output of the command:

	sudo dmsetup table --concise /dev/mapper/lroot

I refactored the code for this version, instead of pretending to be
userspace and performing ioctls, the code just parse the cmd line
argument and create the device directly, which simplified and reduced a
lot the code (and it is much easier to read).

Also, not all dm targets are allowed. Only the ones that I tested were
allowed and the ones that doesn't change any block device when the dm is
create as read-only, i.e. mirror and cache are not allowed, because if
user make a mistake and chose the wrong device to be the mirror or
choose the wrong device to be the cache, it can corrupt data.

So the only targets allowed are:
* crypt
* delay
* linear
* snapshot-origin
* striped
* verity

I wrote a script to perform several tests using qemu that can be found at
https://gitlab.collabora.com/koike/dm-cmdline-test

And you can see the result of the tests at:
https://people.collabora.com/~koike/dm-test-module-param-logs.txt
At the end of this file:
"Total successes: 52. Total failures: 0"

And the output of qemu for each run:
https://people.collabora.com/~koike/dm-test-module-param-logs.tar.gz

I think it would nice to integrate these tests in some CI (maybe kernel
CI?)

Please let me know your comments.
END